rename process net and process client messages functions
